+++ packages/playground/css/async.js
@@ -2,5 +2,7 @@ import './async.css'
 
 const div = document.createElement('div')
 div.className = 'async'
-div.textContent = 'async chunk (this should be teal)'
 document.body.appendChild(div)
+div.textContent = `async chunk (this should be teal) ${
+  getComputedStyle(div).color
+}`