+++ packages/playground/dynamic-import/views/baz.js
@@ -1,1 +1,4 @@
+import { n } from '../nested/shared'
+console.log('baz' + n)
+
 export const msg = 'Baz view'