+++ packages/playground/glob-import/dir/nested/bar.js
@@ -1,4 +1,4 @@
-import modules from 'glob:../*.json'
+const modules = import.meta.globEager('../*.json')
 
 export const msg = 'bar'
 export { modules }