+++ packages/vite/src/node/plugins/dynamicImportPolyfill.ts
@@ -1,16 +1,15 @@
-import MagicString from 'magic-string'
 import { ResolvedConfig } from '..'
 import { Plugin } from '../plugin'
+import { isModernFlag } from './importAnaysisBuild'
 
 export const polyfillId = 'vite/dynamic-import-polyfill'
-const polyfillPlaceholder = `__DYANMIC_IMPORT_POLYFILL__()`
 
 export function dynamicImportPolyfillPlugin(config: ResolvedConfig): Plugin {
   const skip = config.command === 'serve' || config.build.ssr
   let polyfillLoaded = false
   const polyfillString =
-    `${__dynamic_import_polyfill__.toString()};` +
-    `__dynamic_import_polyfill__(${JSON.stringify(config.build.base)});`
+    `const p = ${polyfill.toString()};` +
+    `${isModernFlag}&&p(${JSON.stringify(config.build.base)});`
 
   return {
     name: 'vite:dynamic-import-polyfill',
@@ -27,7 +26,7 @@ export function dynamicImportPolyfillPlugin(config: ResolvedConfig): Plugin {
         polyfillLoaded = true
         // return a placeholder here and defer the injection to renderChunk
         // so that we can selectively skip the injection based on output format
-        return polyfillPlaceholder
+        return polyfillString
       }
     },
 
@@ -43,38 +42,9 @@ export function dynamicImportPolyfillPlugin(config: ResolvedConfig): Plugin {
             `your custom entry.`
         )
       }
-      return {
-        left: '__import__(',
-        right: ')'
-      }
-    },
-
-    renderChunk(code, chunk, { format }) {
-      if (skip) {
-        return null
-      }
-      if (!chunk.facadeModuleId) {
-        return null
-      }
-      const i = code.indexOf(polyfillPlaceholder)
-      if (i < 0) {
-        return null
-      }
-      // if format is not ES, simply empty it
-      const replacement = format === 'es' ? polyfillString : ''
-      if (config.build.sourcemap) {
-        const s = new MagicString(code)
-        s.overwrite(i, i + polyfillPlaceholder.length, replacement)
-        return {
-          code: s.toString(),
-          map: s.generateMap({ hires: true })
-        }
-      } else {
-        return {
-          code: code.replace(polyfillPlaceholder, replacement),
-          map: null
-        }
-      }
+      // we do not actually return anything here because rewriting here would
+      // make it impossible to use es-module-lexer on the rendered chunks, which
+      // we need for import graph optimization in ./importAnalysisBuild.
     }
   }
 }
@@ -111,10 +81,7 @@ declare const document: any
 declare const URL: any
 declare const Blob: any
 
-function __dynamic_import_polyfill__(
-  modulePath = '.',
-  importFunctionName = '__import__'
-) {
+function polyfill(modulePath = '.', importFunctionName = '__import__') {
   try {
     self[importFunctionName] = new Function('u', `return import(u)`)
   } catch (error) {