+++ packages/vite/src/node/server/middlewares/static.ts
@@ -16,11 +16,6 @@ export function serveStaticMiddleware(
   return (req, res, next) => {
     let url = req.url!
 
-    // skip import request
-    if (isImportRequest(url)) {
-      return next()
-    }
-
     // only serve the file if it's not an html request
     // so that html requests can fallthrough to our html middleware for
     // special processing
@@ -30,7 +25,12 @@ export function serveStaticMiddleware(
     ) {
       return next()
     }
-    
+
+    // skip import request
+    if (isImportRequest(url)) {
+      return next()
+    }
+
     // #1426
     url = req.url = decodeURIComponent(url)